Add new Collection.findInStore and Collection.findOneInStore methods to query the store during SSR and use it in posts views
