+++ packages/nova-posts/lib/views.js
@@ -115,8 +115,9 @@ Posts.views.add("userPosts", function (terms) {
 /**
  * @summary User upvoted posts view
  */
-Posts.views.add("userUpvotedPosts", function (terms) {
-  var user = Users.findOne(terms.userId);
+Posts.views.add("userUpvotedPosts", function (terms, apolloClient) {
+  var user = apolloClient ? Users.findOneInStore(apolloClient.store, terms.userId) : Users.findOne(terms.userId);
+
   var postsIds = _.pluck(user.upvotedPosts, "itemId");
   return {
     selector: {_id: {$in: postsIds}, userId: {$ne: terms.userId}}, // exclude own posts
@@ -127,8 +128,9 @@ Posts.views.add("userUpvotedPosts", function (terms) {
 /**
  * @summary User downvoted posts view
  */
-Posts.views.add("userDownvotedPosts", function (terms) {
-  var user = Users.findOne(terms.userId);
+Posts.views.add("userDownvotedPosts", function (terms, apolloClient) {
+  var user = apolloClient ? Users.findOneInStore(apolloClient.store, terms.userId) : Users.findOne(terms.userId);
+
   var postsIds = _.pluck(user.downvotedPosts, "itemId");
   // TODO: sort based on votedAt timestamp and not postedAt, if possible
   return {
